+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Handle `react-router-dom` deprecation warnings. Refs STCOR-448.
 * Update `react-intl` to `v5`. Refs STCOR-449.
 * Add `suppressIntlErrors` option to stripes.config.js.
+* Refactor `CreateResetPassword` to use final-form instead of redux-form. Refs STCOR-441
 
 ## [5.0.2](https://github.com/folio-org/stripes-core/tree/v5.0.2) (2020-06-12)
 [Full Changelog](https://github.com/folio-org/stripes-core/compare/v5.0.1...v5.0.2)